+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Text Input Field
 - FormError Component
 - Form Label Component
+- Checkboxes Component
 
 ### Changed
 
@@ -80,3 +81,11 @@ Initial release of the Design System Package
 ### Added
 
 - Updated code base with "ds-" prefix
+
+## [v1.22.0] - 2022-01-20
+
+### Added
+
+- Checkbox component
+- Form Error Component
+- Form Label Component